requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.6
langchain-groq>=0.0.2
openai>=1.16.1
httpx>=0.27.0
faiss-cpu>=1.7.4
langchain-community>=0.0.30
langchain-core>=0.1.35
langchain-text-splitters>=0.0.1
tiktoken>=0.5.1
requests>=2.31.0